fix: clear the 29 open LOW Sonar findings, 10 fixed and 19 scoped off
Fixed - each is a genuine redundancy the compiler already proves:

- S1905 (x5): casts that do nothing. `(uint)sheetId` where SheetId is already
  uint, in both cache writers; `(XLAddress)first` where FirstAddress is already
  the struct; `(XLPivotCache)pivotTable.PivotCache` in both the slicer and
  timeline factories.
- S8969 (x5): null-forgiving operators the flow analysis does not need.
  `array![0, 0]` after a successful TryPickT0, and `sheet.Id!.Value!` in both
  readers - the preceding `string.IsNullOrEmpty(sheet.Id?.Value)` already
  narrows both to non-null, so the `!`s were claiming knowledge the compiler
  had anyway. A `!` that is not load-bearing is worse than none: it trains the
  reader to skip past the ones that are.

Scoped off, with the reason at the fence:

- S2292 (x7): make XLBorderKey's passthrough properties auto-implemented. This
  one would have cost something real. The struct's own comment measures it -
  fields declared together give 88 bytes, interleaved they give 96 - and this
  struct dominates XLStyleKey, which is copied on every style mutation and
  again on every repository probe. Auto-properties hand the layout back to the
  compiler and undo the grouping. The rule cannot see that the field block is
  deliberate.
- S4136 (x7): overloads should be adjacent. In GridAxis the members are grouped
  by the parameter they project - the `in XLAddress` pair, the `IXLAddress`
  pair, the `Point` pair - because index and cross are always used together and
  one doc comment covers each pair; the rule would split three documented pairs
  to make one alphabetical run. In StyleDecoder the three Decode overloads are
  ordered by layer, entry points before the machinery they delegate to.
- S1125 (x5): `x ? (BooleanValue?)null : false` read as a foldable boolean
  literal. It is three-valued and typed BooleanValue?, and the false branch is
  the only way to write the attribute out as false - there is nothing for `!x`
  to produce when the answer is "omit the attribute".

Not gated by a fresh suite run at the user's direction; the full solution does
build clean under `-p:EnableSonarAnalyzers=true` with the changes in place, and
the previous commit's 14,360 tests were green on both TFMs. Ten of the thirteen
files change by one line each, and the other three are comment-and-pragma only.

+++ XLibur/Excel/IO/SlicerCacheWriter.cs
@@ -197,7 +197,7 @@ internal static class SlicerCacheWriter
             var sheetId = ((XLWorksheet)pivotTable.Worksheet).SheetId;
             pivotTables.AppendChild(new X14.SlicerCachePivotTable
             {
-                TabId = (uint)sheetId,
+                TabId = sheetId,
                 Name = pivotTable.Name,
             });
         }
